Use ssh-keyscan to add ela jump server to known_hosts (#11)
The `AutoAddPolicy` works for the end server which is `daint.cscs.ch` or `eiger.cscs.ch`. For the proxy jump server `ela.cscs.ch`, it is helpful to add it to know host by `ssh-keyscan` so the fingerprint is not asked. This feature is required by the setup of the CSCS machine over a new computational resource widget.